flowernet-web/requirements.txt
fastapi==0.115.13
uvicorn==0.34.0
requests==2.32.3
python-docx==1.1.2
